IVGCVSW-3021 Add end-to-end flow control integration test
Currently asserts that the net fails to optimise as that
is the expected behaviour, but it's complete enough to
exercise most of the code in SwitchLayer.cpp and MergeLayer.cpp

Also, fix a bug in SwitchLayer::ValidateTensorShapesFromInputs
found by the new test.

Also, make topological sort slightly more robust to missing connections
as it should not be the job of the sorter to validate the graph.

Change-Id: I30b9e2d4769ab14a6820284871a79a5bb3eef1ef
Signed-off-by: Matthew Bentham <matthew.bentham@arm.com>