NNXSW-1826 OptimizeConsecutiveReshapes: remove unnecessary call to MoveAllConnections
This is called at a time when newReshape has nothing connected to
its output slot (as it has just been created) and so is a no-op.

The code comment indicated that the intention was to connect the newReshape
to its *input*, but that has already been done in the InsertNewLayer() call
above, so the comment was incorrect.

There is a unit test covering this case ("OptimizeConsecutiveReshapesTest")

Change-Id: I933d5d1c6eb32f5a8269fb5d7c809cd7c89680d1
Signed-off-by: Rob Hughes <robert.hughes@arm.com>