Replace Schema.s with generated .cpp file
This removes the dependency on an assembler (which must support .incbin).
Instead we use xxd as a custom build command to generate a .cpp file
which is included in the build.

This makes it possible to build on Windows, for which you need 
cp.exe (from GNU Core Utils) and xxd.exe (from Vim).

Change-Id: I3a929640b1fdbfc27e2c38abd05bf3a298245ba5
Signed-off-by: Robert Hughes <robert.hughes@arm.com>