Fix TransformIterator operator=
These two operator overloads were giving compile errors as they had
misspelled the names of the member variables.

Presumably they are not used in any of the code and not tested anywhere,
and so the compilers being used on the CI decided to not report this
error. I encountered this error using the MSVC compiler, but it may
cause problems on others too.

Change-Id: I200fc98dbd024d13dbf0e2eca2a0dcd2b0bef261
Signed-off-by: Rob Hughes <robert.hughes@arm.com>