Remove unnecessary friend function
IOptimizedNetwork declares a friend function (CreatePreCompiledWorkloadTest)
that is not necessary as that function does not use any of the non-public
members of the class. I suspect this is leftover after a refactor,
as the function in question uses GetGraphForTesting() to access the
internals instead.

This friend declaration was causing me a compilation error as the namespace
of the declared friend function did not match the namespace of the definition.
Removing it solves the compilation error and tidies up the code.

Change-Id: I4aa893cb5108cd3b72ca355f725fe7a0cd053a7e
Signed-off-by: Rob Hughes <robert.hughes@arm.com>