Fix undefined behaviour in PolymorphicDowncast
In the assertion part of PolymorphicDowncast (and PolymorphicPointerDowncast)
there was an unnecessary static_cast of the source pointer to the destination
type. This is unnecessary because the implicit conversion of the result pointer
back to the base type is sufficient to check for correctness of the downcast,
and potentially harmful because if the downcast is actually incorrect the
behaviour of static_cast is undefined (not actually known to be a problem
in any tested implementation).

Fixes warnings detected by ubsan.

Signed-off-by: Matthew Bentham <matthew.bentham@arm.com>
Change-Id: I9126288cbb06564009f94e57f6ca4688fc3b53c4