Avoid undefined behaviour in NumericCast unit test
Because the maximum value of an int32_t isn't exactly representable
by a 32-bit float, casting it to a float produces a greater value,
and then casting that back to an int32_t is undefined. (In tested
implementations the cast saturates, so accidentally the correct
result is returned).

Using double for the intermediate value instead gives a high
enough precision that this test case can be safely expressed.

Silences warning from Undefioned Behaviour Sanitiser.

Signed-off-by: Matthew Bentham <matthew.bentham@arm.com>
Change-Id: I6fe3e89cb6ac695049c92f3b75aca0c14d0e14aa