Fix compile error in TransformIterator
The operator= functions were raising a compiler error on MSVC.

The implementations of these functions also looked very suspicious as
they were not behaving as a normal operator= would - they did not modify
the 'this' object at all, and returned a copy rather than a reference.

This patch changes them to be more conventional, which required removing
the 'const' qualifier from m_fn in order to make the object assignable.

Change-Id: I1c8b526581ed8973e64f1cffc43b97ac88e07699
Signed-off-by: Rob Hughes <robert.hughes@arm.com>