Fix some Thread Sanitizer warnings
In ConnectionAcknowledgedCommandHandler use std::atomic to manage access to 'enabled' flag.

In StridedSliceAsyncEndToEndTest use separate output buffers for each inference.

Neither of these fixes were likely to be causing real bugs in practice but it's helpful
for Thread Sanitizer to run cleanly so that they don't hide other real bugs.

Signed-off-by: Matthew Bentham <matthew.bentham@arm.com>
Change-Id: I2fcc19bd0afcea3494d9081226754b906bd2bd44