IVGCVSW-6739 Fix logging bug where blank messages were being sent
The ScopedRecord class sends its message to its sinks in its destructor.
The StartNewRecord() method was creating a local ScopedRecord variable,
returning a (move-initialised) copy of this, and then destroying this
local variable (as the variable goes out of scope at the end of the
function). That destruction was causing an empty log message to be sent
to the sinks, which was not intended.

This patch fixes this by making the move constructor of ScopedRecord
disable the moved-from object, so that it does not send its message
upon destruction.

Change-Id: I435d96074698575ed3445fed2597c115b83701fd
Signed-off-by: Rob Hughes <robert.hughes@arm.com>